+++ dart/lib/src/protocol.dart
@@ -1,0 +1,13 @@
+export 'protocol/app.dart';
+export 'protocol/breadcrumb.dart';
+export 'protocol/browser.dart';
+export 'protocol/contexts.dart';
+export 'protocol/device.dart';
+export 'protocol/dsn.dart';
+export 'protocol/event.dart';
+export 'protocol/level.dart';
+export 'protocol/package.dart';
+export 'protocol/runtime.dart';
+export 'protocol/sdk.dart';
+export 'protocol/system.dart';
+export 'protocol/user.dart';